fix: correct whale_gate's min_contracts type and scope mypy's import-following
Two real gaps found by the quality-architecture-audit CI pipeline (which
runs a Kalshi-boundary-scoped mypy over services/kalshi/) on this
branch's first push:

1. whale_gate.passes()'s min_contracts parameter was typed int, but the
   function it wraps (kalshi_trade_tape.min_contracts_for, reused per
   this plan's own Task 3 instruction) legitimately returns float - a
   real type mismatch, harmless at runtime (Python's numeric comparison
   doesn't care) but exactly what this static check exists to catch
   before it isn't harmless somewhere else.

2. Task 3's reuse of kalshi_trade_tape.min_contracts_for pulled
   config_store.py, config_bounds.py, and services/market_analyst_agent/*
   into mypy's import graph for the first time - none of them previously
   reachable from mypy.ini's `files = services/kalshi, services/execution.py`
   scope, and none of them touched by this plan. mypy's default
   import-following reported 18 pre-existing type errors in those
   unrelated legacy modules as if new. `follow_imports = silent` keeps
   inference accurate through the reused function while restoring the
   file's own stated intent ("Deliberately NOT repo-wide... without
   failing the legacy application on unrelated typing debt").

+++ services/whale_gate.py
@@ -11,7 +11,7 @@ from services.kalshi.contracts import trade as trade_contract
 from services.whalewatchers.kalshi_trade_tape import min_contracts_for as _series_min_contracts_for
 
 
-def passes(trade: dict, *, min_contracts: int) -> bool:
+def passes(trade: dict, *, min_contracts: float) -> bool:
     """True if this raw trade print's contract count clears min_contracts.
     A trade with an unparseable/missing count fails open toward the caller
     (returns False here, but callers must fall open on an *exception*, not
